separate v4 rules
