+++ package.json
@@ -7,7 +7,8 @@
     "test": "mocha packages/eslint-plugin-pf-codemods/test/rules/*",
     "test:v4": "mocha packages/eslint-plugin-pf-codemods/test/rules/v4",
     "test:v5": "mocha packages/eslint-plugin-pf-codemods/test/rules/v5",
-    "test:single": "pf-codemods --no-cache test/test.tsx",
+    "test:v4:single": "pf-codemods --v4 --no-cache test/v4test.tsx",
+    "test:v5:single": "pf-codemods --no-cache test/test.tsx",
     "test:koku": "pf-codemods --no-cache test/koku-ui",
     "test:console": "pf-codemods --no-cache test/console/frontend",
     "test:integreatly": "pf-codemods --no-cache test/tutorial-web-app"
