+++ packages/eslint-plugin-pf-codemods/index.js
@@ -1,5 +1,32 @@
-const rules = {
+const v5rules = {
   "accordion-rename-displaySize-large": require('./lib/rules/v5/accordion-rename-displaySize-large'),
+  "card-remove-isHoverable": require('./lib/rules/v5/card-remove-isHoverable'),
+  "charts-resizeObserver-import": require('./lib/rules/v5/charts-resizeObserver-import'),
+  "clipboardCopy-remove-popoverPosition": require('./lib/rules/v5/clipboardCopy-remove-popoverPosition'),
+  "datalist-remove-ondrags": require('./lib/rules/v5/datalist-remove-ondrags'),
+  "divider-remove-isVertical": require('./lib/rules/v5/divider-remove-isVertical'),
+  "expandable-section-rename-displaySize-large": require('./lib/rules/v5/expandable-section-rename-displaySize-large'),
+  "fileUpload-remove-onChange": require('./lib/rules/v5/fileUpload-remove-onChange'),
+  "horizontalSubnav-ariaLabel": require('./lib/rules/v5/horizontalSubnav-ariaLabel'),
+  "pagination-optionsToggle": require('./lib/rules/v5/pagination-optionsToggle'),
+  "pagination-remove-ToggleTemplateProps": require('./lib/rules/v5/pagination-remove-ToggleTemplateProps'),
+  "pagination-rename-props": require('./lib/rules/v5/pagination-rename-props'),
+  "remove-sticky-props": require('./lib/rules/v5/remove-sticky-props'),
+  "resizeObserver-function-param": require('./lib/rules/v5/resizeObserver-function-param'),
+  "simpleList-remove-isCurrent": require('./lib/rules/v5/simpleList-remove-isCurrent'),
+  "spinner-svg-default": require('./lib/rules/v5/spinner-svg-default'),
+  "tableComposable-remove-hasSelectableRowCaption": require('./lib/rules/v5/tableComposable-remove-hasSelectableRowCaption'),
+  "tabs-rename-hasBorderBottom": require('./lib/rules/v5/tabs-rename-hasBorderBottom'),
+  "tabs-remove-hasSecondaryBorderBottom": require('./lib/rules/v5/tabs-remove-hasSecondaryBorderBottom'),
+  "toggle-remove-isPrimary": require('./lib/rules/v5/toggle-remove-isPrimary'),
+  "toolbar-remove-visiblity": require('./lib/rules/v5/toolbar-remove-visiblity'),
+  "tooltip-remove-props": require('./lib/rules/v5/tooltip-remove-props'),
+  "remove-removeFindDomNode": require('./lib/rules/v5/remove-removeFindDomNode'),
+  "popover-remove-props": require('./lib/rules/v5/popover-remove-props'),
+  "warn-key-codes-removed": require('./lib/rules/v5/warn-key-codes-removed'),
+};
+
+const v4rules = {
   "accordion-remove-noBoxShadow": require('./lib/rules/v4/accordion-remove-noBoxShadow'),
   "application-launcher-rename-dropdownItems": require('./lib/rules/v4/application-launcher-rename-dropdownItems'),
   "aria-props": require('./lib/rules/v4/aria-props'),
@@ -18,7 +45,6 @@ const rules = {
   "background-image-src-enum": require('./lib/rules/v4/background-image-src-enum'),
   "skip-to-content-remove-component": require('./lib/rules/v4/skip-to-content-remove-component'),
   "alert-new-action": require('./lib/rules/v4/alert-new-action'),
-  "card-remove-isHoverable": require('./lib/rules/v5/card-remove-isHoverable'),
   "card-rename-components": require('./lib/rules/v4/card-rename-components'),
   "chipgroup-remove-props": require('./lib/rules/v4/chipgroup-remove-props'),
   "chipgroup-remove-chipbutton": require('./lib/rules/v4/chipgroup-remove-chipbutton'),
@@ -28,7 +54,6 @@ const rules = {
   "empty-state-icon-removed-props": require('./lib/rules/v4/empty-state-icon-removed-props'),
   "pagination-removed-variant": require('./lib/rules/v4/pagination-removed-variant'),
   "remove-isPseudo-props": require('./lib/rules/v4/remove-isPseudo-props'),
-  "remove-sticky-props": require('./lib/rules/v5/remove-sticky-props'),
   "label-remove-isCompact": require('./lib/rules/v4/label-remove-isCompact'),
   "rename-noPadding": require('./lib/rules/v4/rename-noPadding'),
   "tab-title-text": require('./lib/rules/v4/tab-title-text'),
@@ -38,7 +63,6 @@ const rules = {
   "no-unused-imports": require('./lib/rules/v4/no-unused-imports'),
   "tab-rename-variant": require('./lib/rules/v4/tab-rename-variant'),
   "form-fix-isValid": require('./lib/rules/v4/form-fix-isValid'),
-  "expandable-section-rename-displaySize-large": require('./lib/rules/v5/expandable-section-rename-displaySize-large'),
   "expandable-rename-expandablesection": require('./lib/rules/v4/expandable-rename-expandablesection'),
   "rename-toolbar-components": require('./lib/rules/v4/rename-toolbar-components'),
   "page-header-prop-rename": require('./lib/rules/v4/page-header-prop-rename'),
@@ -48,31 +72,20 @@ const rules = {
   "chartVoronoiContainer-remove-allowTooltip": require('./lib/rules/v4/chartVoronoiContainer-remove-allowTooltip'),
   "chart-remove-allowZoom": require('./lib/rules/v4/chartVoronoiContainer-remove-allowTooltip'),
   "react-icons-remove-icon": require('./lib/rules/v4/react-icons-remove-icon'),
-  "charts-resizeObserver-import": require('./lib/rules/v5/charts-resizeObserver-import'),
-  "clipboardCopy-remove-popoverPosition": require('./lib/rules/v5/clipboardCopy-remove-popoverPosition'),
-  "datalist-remove-ondrags": require('./lib/rules/v5/datalist-remove-ondrags'),
-  "divider-remove-isVertical": require('./lib/rules/v5/divider-remove-isVertical'),
-  "fileUpload-remove-onChange": require('./lib/rules/v5/fileUpload-remove-onChange'),
-  "horizontalSubnav-ariaLabel": require('./lib/rules/v5/horizontalSubnav-ariaLabel'),
-  "pagination-optionsToggle": require('./lib/rules/v5/pagination-optionsToggle'),
-  "pagination-remove-ToggleTemplateProps": require('./lib/rules/v5/pagination-remove-ToggleTemplateProps'),
-  "pagination-rename-props": require('./lib/rules/v5/pagination-rename-props'),
-  "resizeObserver-function-param": require('./lib/rules/v5/resizeObserver-function-param'),
-  "simpleList-remove-isCurrent": require('./lib/rules/v5/simpleList-remove-isCurrent'),
-  "spinner-svg-default": require('./lib/rules/v5/spinner-svg-default'),
-  "tableComposable-remove-hasSelectableRowCaption": require('./lib/rules/v5/tableComposable-remove-hasSelectableRowCaption'),
-  "tabs-rename-hasBorderBottom": require('./lib/rules/v5/tabs-rename-hasBorderBottom'),
-  "tabs-remove-hasSecondaryBorderBottom": require('./lib/rules/v5/tabs-remove-hasSecondaryBorderBottom'),
-  "toggle-remove-isPrimary": require('./lib/rules/v5/toggle-remove-isPrimary'),
-  "toolbar-remove-visiblity": require('./lib/rules/v5/toolbar-remove-visiblity'),
-  "tooltip-remove-props": require('./lib/rules/v5/tooltip-remove-props'),
-  "remove-removeFindDomNode": require('./lib/rules/v5/remove-removeFindDomNode'),
-  "popover-remove-props": require('./lib/rules/v5/popover-remove-props'),
-  "warn-key-codes-removed": require('./lib/rules/v5/warn-key-codes-removed'),
 };
 
 // if you want a rule to have a severity that defaults to warning rather than error, add the rule name to the below array
 const warningRules = ["horizontalSubnav-ariaLabel"]
+
+const createRules = (rules) => {
+  return Object.keys(rules).reduce((acc, rule) => {
+    const severity = warningRules.includes(rule) ? "warn" : "error"
+    acc[`@patternfly/pf-codemods/${rule}`] = severity;
+    return acc;
+  }, {})
+}
+
+const mappedRules = {...createRules(v5rules), ...createRules(v4rules)};
 
 module.exports = {
   configs: {
@@ -92,12 +105,9 @@ module.exports = {
           jsx: true,
         },
       },
-      rules: Object.keys(rules).reduce((acc, rule) => {
-        const severity = warningRules.includes(rule) ? "warn" : "error"
-        acc[`@patternfly/pf-codemods/${rule}`] = severity;
-        return acc;
-      }, {}),
+      rules: mappedRules,
     },
   },
-  rules,
+  rules: {...v5rules, ...v4rules},
+  ruleVersionMapping: {"v4": Object.keys(v4rules), "v5": Object.keys(v5rules)}
 };
